+++ app/Repository/CocktailRepository.php
@@ -8,7 +8,6 @@ use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\DB;
 use Kami\Cocktail\Models\Cocktail;
 use Illuminate\Database\DatabaseManager;
-use Kami\Cocktail\Models\ComplexIngredient;
 
 readonly class CocktailRepository
 {
@@ -26,48 +25,47 @@ readonly class CocktailRepository
     public function getCocktailsByIngredients(array $ingredientIds, ?int $limit = null, bool $useParentIngredientAsSubstitute = false, bool $matchComplexIngredients = true): Collection
     {
         // Resolve complex ingredients
+        // Basically, goes through all ingredients to match ($ingredientIds) and check if they can create complex ingredients
+        // If they can, that ingredient is added to the list of ingredients to match
         if ($matchComplexIngredients) {
-            $complexIngredientsSubquery = ComplexIngredient::select('main_ingredient_id', DB::raw('COUNT(DISTINCT ingredient_id) as cnt'))
-                ->whereIn('ingredient_id', $ingredientIds)
-                ->groupBy('main_ingredient_id');
+            $rawQuery = "WITH RECURSIVE IngredientChain AS (
+                    SELECT id AS matched_ingredient
+                    FROM ingredients
+                    WHERE id IN (" . implode(',', $ingredientIds) . ")
+                    UNION
+                    SELECT ci.main_ingredient_id AS matched_ingredient
+                    FROM complex_ingredients ci
+                    INNER JOIN IngredientChain ic ON ci.ingredient_id = ic.matched_ingredient
+                )
+                SELECT DISTINCT matched_ingredient
+                FROM IngredientChain;";
 
-            $matchedComplexMainIngredientIds = DB::table(DB::raw("({$complexIngredientsSubquery->toSql()}) as IngredientCount"))
-                ->mergeBindings($complexIngredientsSubquery->getQuery())
-                ->where('cnt', count($ingredientIds))
-                ->pluck('main_ingredient_id');
-
-            foreach ($matchedComplexMainIngredientIds as $additionalId) {
-                if (!in_array($additionalId, $ingredientIds)) {
-                    $ingredientIds[] = $additionalId;
-                }
-            }
+            $additionalIngredients = collect(DB::select($rawQuery))->pluck('matched_ingredient');
+            $ingredientIds = array_merge($ingredientIds, $additionalIngredients->toArray());
+            $ingredientIds = array_unique($ingredientIds);
         }
 
         $query = $this->db->table('cocktails AS c')
             ->select('c.id')
             ->join('cocktail_ingredients AS ci', 'ci.cocktail_id', '=', 'c.id')
+            ->join('ingredients AS i', 'i.id', '=', 'ci.ingredient_id')
             ->leftJoin('cocktail_ingredient_substitutes AS cis', 'cis.cocktail_ingredient_id', '=', 'ci.id')
-            ->where('optional', false);
+            ->where('optional', false)
+            ->where(function ($query) use ($ingredientIds, $useParentIngredientAsSubstitute) {
+                $query->whereIn('i.id', $ingredientIds)->orWhereIn('cis.ingredient_id', $ingredientIds);
 
-        if ($useParentIngredientAsSubstitute) {
-            $query->join('ingredients AS i', function ($join) {
-                $join->on('i.id', '=', 'ci.ingredient_id');
+                if ($useParentIngredientAsSubstitute) {
+                    $query->orWhereIn('i.id', function ($parentSubquery) use ($ingredientIds) {
+                        $parentSubquery
+                            ->select('parent_ingredient_id')
+                            ->from('ingredients')
+                            ->whereIn('id', $ingredientIds)
+                            ->whereNotNull('parent_ingredient_id');
+                    });
+                }
             })
-            ->leftJoin('ingredients AS pi', function ($join) { // Faster than OR join
-                $join->on('pi.parent_ingredient_id', '=', 'ci.ingredient_id');
-            })
-            ->whereIn('i.id', $ingredientIds)
-            ->orWhereIn('i.parent_ingredient_id', $ingredientIds)
-            ->orWhereIn('pi.id', $ingredientIds)
-            ->orWhereIn('pi.parent_ingredient_id', $ingredientIds);
-        } else {
-            $query->join('ingredients AS i', 'i.id', '=', 'ci.ingredient_id')
-            ->whereIn('i.id', $ingredientIds);
-        }
-
-        $query->orWhereIn('cis.ingredient_id', $ingredientIds)
-        ->groupBy('c.id')
-        ->havingRaw('COUNT(*) >= (SELECT COUNT(*) FROM cocktail_ingredients WHERE cocktail_id = c.id AND optional = false)');
+            ->groupBy('c.id')
+            ->havingRaw('COUNT(*) >= (SELECT COUNT(*) FROM cocktail_ingredients WHERE cocktail_id = c.id AND optional = false)');
 
         if ($limit) {
             $query->limit($limit);